backend/requirements.txt
Flask
flask-cors
pdfplumber
faiss-cpu
sentence-transformers
requests
